[LearnOpenGL] Hello Triangle
Following the tutorial through again, took forever to get it working
because I missed binding the vertex array like a chump.

Did a little refactoring to make it easier to see what the stages are.

Saved out a screenshot of my triangle because I was very proud when I
finally got it working :)