Allow between gouraud and phong shading
Keyboard key 3 will now toggle between shading styles.
Made use of shared uniforms by using the precision qualifier "highp" so
that uniforms only need to be sent once and can be used by both vertex
and fragment shaders.
Booleans can't use the highp precision qualifier, so it is sent as an
int and converted back to a bool inside the shader program.