Add support for multiple light sources
Refactored fragment shader to use functions. Lights are now expected to
come in as a list of structs, with the type attribute defining how it
will be processed.

Refactored World class to handle a list of lights.
Moved model calculation to the World::updateModel function, and set a
variable on each Shader object to hold it's unique model matrix.

Commented out some keyboard input functions from utils, because they
need re-implementing to handle multiple which I didn't want to focus on
yet. I might re-implement them, but some thing like shading in view
space were just exercised left over and were making things unnecessarily
complicated so I took them out permamently.

Added in a lot more logging to help debugging, which I finally
discovered was because normal matrix was being generated as a mat4 but
the uniform needed to be a mat3 and it was silently failing.