Debugging mesh and model - WIP 3
Managed to get the model loading and the geometry seems to be OK, but
the textures are very much not OK.
Almost looks like the UVs are not mapped correctly, so it might be
something to do with flipping (or not flipping) the textures.

Spotlight is blacking out instead of illuminating its cone of influence.
Probably need to set up some static lights as well as randomised lights
are current inside the model, which might also cause normal issues.

Fixed a bunch of bugs;
 - processMesh wasn't returning a value
 - string comparison for textures wasn't working
 - already loaded textures weren't being cached
 - Removed shader from mesh and model draw functions, as not necessary
   and was causing circular imports
 - Added a default constructor for Model so it could be an attribute of
   the shader
 - Mesh deconstructor was deleting buffers when not necessary