Fixed lighting issue at glancing angles
Blackouts were being caused by shininess not being set when loading from
a model, which was usually set in the Shader::loadMaterials function.
I should be able to get the shininess from the aiMaterial, but getting
this to be used per mesh might require some more refactoring.

Thought it might have been the variable 'distance' overriding a function
name in the fragment shader, but that didn't seem to have any effect.
Kept the change anyway, since it's not such a good idea to override
function names with variables.

Also renamed the texture types to be used on Texture class
initialisation since it wasn't being used before.

TODO:
 - extract shininess dynamically from loaded materials
 - set shininess uniform per mesh draw (if necessary) or per shader.
 - convert lights array in main to a std::vector