Add ability to move selected light
Currently only able to move forward and back, and the light source's
shader is not picking up the change in the geometry, but the light
position inside the object shader is changing.

Camera can still be moved about as before, but now the user needs to
hold down the space bar.