Render outline on selected light
Currently working for the original selection, inlcuding if the light is
moved, but for some reason the stencil mask does not work on other
lights if they are selected with tab. The outline pass doesn't seem to
pick up the stencil and just fills in the entire shape with outline
colour.