Add transparancy render pass by distance
Opaque meshes are now collected on intialisation and then saved for a
separate pass after all other objects have been rendered, where they are
rendered in reverse order based on distance from camera.

For some reason logging won't turn off now inside the meshes after the
first pass through, so I have disabled it.

Also the outline pass seems to be intefering with the opacity pass, so
either some tweaking needs to be done to the stencil, or the opacity
pass order moved after the stencil, or both.