Add skybox
Skybox geometry is dynamically generated when Shader is instantiated
with a new 'CubeMap' texture class.

Few issues with this,
 - the cubemap texture is upside down
 - only one of the 6 cubemap faces appears to be visible
 - cubemap appears to be rendering on top of everything else instead of
   behind as was the plan.

I think there might be some issues because the cube is generated
positive Z first, and I think it needs to be positive X first to match
the cubemap loading pattern.

Not sure why it's rendering on top though, will have to investigate.