Fix skybox covering objects
Rendering the skybox last, and swizzling the skybox z value fixes the
issue where the skybox was covering some objects.

However, the outlines around highlighted lights still seems to be
broken, so I will have to figure that out.

Also the skybox is still flipped weirdly, so I'll have to find out
what's going on there. Seems like it's something to do with the
coordinate system. I want to try to fix this in code, rather than
messing around with the skybox textures to get a better understanding of
the coordinate systems as it's confusing me a lot.