Fix skybox orientation
The issue was caused by cubemaps apparently using a left handed
coordinate system, not the right handed coord system used in the rest of
opengl. Flipping in Y was no good because it makes the sky on the floor,
and doesn't doesn't solve flipping in X and Z anyway.

In the end I swapped the load order of the top and bottom using the
cubemap enums, and then in the vertex shader I made the negative
position the texture coordinate, meaning it's flipped in X, Y and Z
which seems to do the trick.

There's probably a better solution for swapping between coordinate
systems, and ultimately I think it's probably down to using input data
set up the coordinate system rather than altering the coordinate system
to match the data... but there you go. I did figure out what was going
on though, so that's nice.

Still losing the outline of the selected light though.