Fix outlines and transparency while using skybox
Issue was caused by the render order of objects. Because the depth mask
is temporarily disabled when rendering the skybox (because in reality
it's a very close 1x1x1 cube, not a very far away one) it passes the
stencil test as well.

Also, didn't notice this before, but the transparency of the lightbulb
was being lost as well because when the bulb was rendered, at that point
there's nothing behind it, so nothing to blend with.

Fixed both of these issues by slotting the skybox render call in between
opaque objects and transparent objects. This makes the final render
order: opaque > skybox > transparent > outlines/stencils.

Included an example render for posterity.