Add refraction
For some reason I had to negate both the incoming vector and the normal
vector to get it to look right. I'm guessing it's something to do with
the coordinate system again, because without it everthing seems to in
the opposite direction.

Also tweaked the reflection to be relative to the material shininess
rather than being a hard coded value.

Also played around with a few different models while I was figuring out
what was going on with refraction because, for example, spheres have an
inverted image under refraction which wasn't helping my figure out why
it was wrong (or if it was even wrong at all).