Set up geometry shader pass-through loading obj
Also set up in/out variables for shader stages as interface blocks which
should make them a bit easier to manage. They interface block is the
same across all stages, which should make it easier to remove the
geometry shader without having to rename everything.

Removed the gouraud shading because we're not using it anymore.