Add normal visualisation
Ended up having to make a whole separate function to generate the
normals shader, because when I tried to get Shader::createShaderProgram
to return an unsigned int so the function could be re-used to create a
second shader program, the whole program stopped rendering anything.

Which is bizarre, because the value was definitely being assigned, and
returning from the function, but for some reason it didn't like it. I
didn't like it either, so at least the feeling was mutual.