Add many rocks without instancing
To prepare for setting up a scene with proper instances, made various
changes to set the scene up with the planet and rocks. It's quite slow
to load because the shader program has to be created for each one
uniquely, but it looks kinda cool! Included a screenshot for posterity.

Added a rotation function to the model matrix calculations to enable
objects to face in their given direction (this also means a direction
must be set of the model matrix zeroes out and the vertices can't be
draw).

I also fixed a bug that had crept in somehow where the Up and
Forward controls were the same. Selected object (light) now moves as
expected along it's local axis.

Finally, converted the list of lights and objects to a list of pointers,
as this is probably more memory efficient. Not gonna lie, I'm just
following how the tutorial had it, I don't 100% understand when data is
copied around (or not) because of pointers.