Set up off-screen multi-sample anti-aliasing
Front and reverse frame buffers are now created as multi-sample
framebuffers, and an intermediary buffer that uses regular textures gets
copied to when drawing to the quads, allowing post-processing to occur
there.

Framebuffer setup is getting a little cumbersome now, so I'm going to
refactor this into a few functions.