Fix normal shader for non-instanced objects
Figured it would be easier to set it up for non-instanced objects first,
as the normal toggle is global, but not all objects may have one. I
should really change this to a check to see if a normal shader is
available, but I don't want to over-engineer it at this point.

Since creating the normal shader is its own function, I can choose what
objects will get a normal shader, so it wouldn't be too hard to
refactor, I just want to move on to other things for now.