Add optional blinn-phong lighting model
User can now switch between blinn and standard phong lighting with
hotkey '2'. I also included a calculaion for energy conservation, which
allegedly compensates for the different shading models by adjusting the
shininess, but I'm not convinced it works for the regular phong model,
because while using it the specular highlights seem far too bright.

Left it in though, because why not.