Add fixed gamma correction colour pipeline
Gamma is now applied at the end in the post processing shaders, with all
textures loaded in to linear colourspace by default.

Works better than before, giving slightly more realistic results, but
had issues with attenuation not doing what I thought it should, so in
the end I just left attenuation as it was.